v.1.0.6
-> Updated caching strategy on sw
-> Added in an event listener on load on root to make sure assets load prior to registering service worker (improves stability, hinders performance)
-> Added sw worker version.
-> Added local and online domain const

:computer:
:coffee:
:cop:
:crescent_moon:

Current behavior:

If you click on a snippet, while offline and it isn't cached -> It routes back to snippets page and app doesn't fall and crash most times.

Expected behavior:

Show a message saying you are offline and the snippet you selected isn't in cache.

// ----------- TODO --------------- //
-> Get the precached offline page to render when user navigates to route or asset that is not in cache and not online.
-> We can futher enhance the offline experience by validating which url we are on and sending an expected response back to the user from the offline cache we make.

+++ app/routes/snippets/seed.jsx
@@ -22,7 +22,7 @@ export async function action({ request }) {
     const userID = session.get("userID");
     // Delete all Snippets on Database if not already empty
     if (snippetCount !== 0) {
-      const deletedManySnippets = await db.models.snip.deleteMany();
+      const deletedManySnippets = await db.models.snip.deleteMany({ user: userID });
       // Inset Default Seed to Database -> Data pumping to add more data
       for (let x = 0; x < 20; x++) {
         snippetSeed.forEach((snippetData, i) => {
